Add Blake3 support to hash_utils
